app/core/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
cssselect==1.3.0
cssutils==2.11.1
fastapi==0.116.1
idna==3.10
lxml==6.0.0
more-itertools==10.7.0
premailer==3.10.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yagmail==0.15.293
